Lab 7

+++ src/main/java/complllex/backend/models/Museum.java
@@ -11,14 +11,14 @@ import java.util.Set;
 public class Museum {
 
     public Museum() { }
-    public Museum(Long id) {
+    public Museum(int id) {
         this.id = id;
     }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", updatable = false, nullable = false)
-    public long id;
+    public int id;
 
     @Column(name = "name", nullable = false, unique = true)
     public String name;
