+++ src/main/java/complllex/backend/repositories/ArtistRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 import complllex.backend.models.Artist;
 
 @Repository
-public interface ArtistRepository  extends JpaRepository<Artist, Long> {
+public interface ArtistRepository  extends JpaRepository<Artist, Integer> {
 
 }
