+++ src/main/java/complllex/backend/repositories/MuseumRepository.java
@@ -3,5 +3,5 @@ package complllex.backend.repositories;
 import complllex.backend.models.Museum;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MuseumRepository extends JpaRepository<Museum, Long> {
+public interface MuseumRepository extends JpaRepository<Museum, Integer> {
 }
